protocol-locators are defined in a single file
