+++ packages/wallet/src/communication/protocol-locator.ts
@@ -1,0 +1,9 @@
+import { makeLocator } from '../redux/protocols';
+import { EmbeddedProtocol } from '.';
+
+export const ADVANCE_CHANNEL_PROTOCOL_LOCATOR = makeLocator(EmbeddedProtocol.AdvanceChannel);
+export const CONSENSUS_UPDATE_PROTOCOL_LOCATOR = makeLocator(EmbeddedProtocol.ConsensusUpdate);
+export const EXISTING_LEDGER_FUNDING_PROTOCOL_LOCATOR = makeLocator(
+  EmbeddedProtocol.ExistingLedgerFunding,
+);
+export const LEDGER_TOP_UP_PROTOCOL_LOCATOR = makeLocator(EmbeddedProtocol.LedgerTopUp);
