+++ packages/wallet/src/redux/protocols/advance-channel/index.ts
@@ -1,3 +1,5 @@
+export { ADVANCE_CHANNEL_PROTOCOL_LOCATOR } from '../../../communication/protocol-locator';
+
 export { AdvanceChannelState, isSuccess } from './states';
 export { AdvanceChannelAction, isAdvanceChannelAction } from './actions';
 export {
